Updated data server to use doubles
Former-commit-id: bbee6f184c249eadb334b7989aea2a71ef0179b2

+++ ros_sim/data_server/include/data_server/DataServerEntry.h
@@ -3,16 +3,16 @@
 
 struct DataServerEntry
 {
-	float x;
-	float y;
-	float h;
-	float sonarDepth;
+	double x;
+	double y;
+	double h;
+	double sonarDepth;
 	ros::Time time;
 
-	float temp;
-	float salt;
-	float dye;
-	float plumeStrength;
+	double temp;
+	double salt;
+	double dye;
+	double plumeStrength;
 
 	bool operator<(const DataServerEntry& rhs) const { return time < rhs.time; }
 	bool operator==(const DataServerEntry& rhs) const { return time == rhs.time; }
